pyproject.toml
[tool.poetry]
name = "repl_python3_dcTestBot233"
version = "0.1.0"
description = ""
authors = ["BinLYing <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
discord = "^1.7.3"
Flask = "^2.1.2"
lxml = "^4.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
